refactor(train): S3a —— 六个 Anima 专属模块 git mv 进 families/anima/（多模型 PR-2b 阶段 3a）
text_encoding / comfy_qwen / sampling / navit / leap / sra_align →
runtime/training/families/anima/（D5 点名的 Anima-only 代码归位）；
全仓 import 重指向（含 from training import X 形式两处漏网）。纯机械移动。
