+++ tests/test_navit_packed_objective.py
@@ -17,7 +17,7 @@ pytest.importorskip("xformers")
 from modeling.anima.anima_modeling import Anima
 from modeling.anima.cosmos_predict2_modeling import set_xformers_enabled
 from training.losses.mse import MseLoss
-from training.navit import navit_packed_forward_and_loss, pack_cross_embeddings
+from training.families.anima.navit import navit_packed_forward_and_loss, pack_cross_embeddings
 
 requires_cuda = pytest.mark.skipif(
     not torch.cuda.is_available(), reason="needs CUDA + xformers"
